Set default region to 'CAN' in visualization scripts
This commit updates the visualization scripts so that the default region value is 'CAN' if it exists in the available regions, otherwise it will fall back to the first region in the list. It also updates the 'main.py' script to run the app in debug mode. This change makes the application more user-friendly for users primarily concerned with Canadian regions and aims to facilitate debugging during development.