Add data loading notification and error handling
A data loading notification has been added to the data modal component to better indicate when data is being processed. Furthermore, error handling has been introduced to the selected_files component to ensure only CSV files are supported and to show an alert when a different file format is attempted to be loaded.