Enhance visualizations with bar and map plot options
This commit refactors the data visualization scripts for 'Transmission Flow' and 'Transmission Capacity' to include options for both Map Plot and Bar Plot. Previously only one type of visualization was available but with this update, the users can choose between the 'Map Plot' and 'Bar Plot' options. This change adopts more user data preferences and enhances the data exploration experience. Moreover, it simplifies the function calls and enhances code organization and manageability.