Add 'Generation Type Data' visualization and processing scripts
Introduced scripts for visualizing and processing 'Generation Type Data' in the copper input profile. This includes the generation of plots for 'Generation Type Data' and the establishment of related callbacks, as well as the processing of this data from the DataFrame. Adaptations were carried out in the copper input profile main script and plots configuration, to incorporate these new functionalities.