Refactored code to include 'region' in data processing and visualization
Changes were made across several processing and visualization scripts to stop restricting the data to Canadian region only and include 'region' as part of the returned data. This allows for broader data analysis and visualization by expanding the scope to include all regions. Moreover, the 'region' selector has been added to the visualization interface to enable the user to choose from the available regions for the display of data.