Refactor profile module handling
Refactored how the profile modules are handled in the 'data_viewer_handling.py' file. Updated the code to use the .get() method for dictionary lookup of profile modules and added a guard to skip in case module key is not found. This change prevents unforeseen errors when a specified profile module doesn't exist in the dictionary, improving the stability of the application.