Update gen_type_copper field aggregation in generation_capacity.py
In the script 'generation_capacity.py', an alteration has been applied to conditionally aggregate the 'gen_type_copper' field data. If the aggregation is false, the values are transformed using 'utils.get_name()' method. This provides the flexibility to breakdown or group the 'gen_type_copper' data depending on the requirements.