Update scripts for capacity, transmission and VRE generation visualization
Updated the scripts to improve code efficiency and functionality for generation capacity and transmission data. Adjustments were made to data categorization and button interactions in the generation capacity script, grouping methods in the processing script, and the render functions and hover templates in the visualization scripts. Also, integrated VRE capacity factor generation with inclusion of geographical representations in maps.