Update vre.py to remove CRS transformations
The changes in this commit remove the coordinate reference system (CRS) transformations that were previously implemented in the vre.py file. The code no longer sets the CRS to "EPSG:4326" and converts it to "EPSG:5070". As a result, the function will now return data in the input's given CRS, without any modifications.