Refactor code and improve data clarity in energy model
Refactored multiple scripts in the energy model by grouping scenarios, filter and format data more effectively. In 'cost_total.py', the check for whether model equals 'PyPSA_CAN' is commented out for now. In 'overview.py', the dataframe is now sorted by time. In other scripts, 'Operational|' and 'Capital|' are tweaked to start with this text to improve understandability of variable type. In 'energy_model.py', the time and period filters in the data now include both integer and string inputs for years for broader coverage. Finally, a new 'Qualifying Capacity' plot was added in 'plots.yaml'.