Add database selection in data runs filtering
A new feature has been added to the input options in the function get_runs in the database_connection.py file. This feature allows the users to select a specific database (DB) from the list of databases when running the function. This will help to narrow down the selection of data runs to a specific database if desired. The id for this new selection option has been added in the ids.py file.