Remove 'change_models.ipynb' file and disable closing in plot window
The 'change_models.ipynb' file is no longer needed and was deleted in this commit. Meanwhile, in 'plot_window/window.py', the option to close the plot window has been turned off by setting 'closable' attribute to false. This change is expected to enhance the user experience, since there is a dash internal bug, that freezes the page after clicking the 'x', workaround is using the trashcan in the sidebar