Add port argument to open_browser function
The previous version of open_browser function was modified to include a port argument. This was done to ensure that the desired port is utilized when starting the server and opening the browser, increasing the code's flexibility. The port is now defined in the main function and passed as an argument when the Timer starts.