Add 'CIMS' to model_mapping and improve column checking
The 'CIMS' model has been added to the model_mapping. The code regarding column checking has also been revised for more accurate results. Now, instead of generally checking for required columns ('model', 'scenario', 'variable', 'value', 'region', 'time'), the code first checks if the 'model' is in df.columns and that it is included in the model_mapping; if not, then it proceeds with the original check. If the model is in df.columns and included in model_mapping, it ensures that 'scenario' column is present.