Add greenhouse gas (GHG) emissions to CIMS output
The functionality of the CIMS output has been expanded to include greenhouse gas emissions. GHG emissions can now be processed, visualized, and interacted with similarly to other provided quantities. This includes functions to check, process, and visualize them and the necessary callbacks to link action and response, enhancing the scope of our model output data.