Update plot rendering and configuration for stock and emissions
This updates the plot rendering function in visualization scripts for both the stock and Greenhouse Gas (GHG) scenarios. A config from a YAML file is used to determine the plot's name, unit, and labels. Furthermore, changes have been implemented in callback and processing scripts to incorporate the plot name. This makes the process of rendering different types of plots more streamlined and consistent.