Add plot settings to visualization scripts
Names and units have been added to all plots for better clarity and understanding. Changes were made in the YAML, and then used in all corresponding visualization scripts ensuring that plot representations are consistent across the application. The processing script for the stock module was also tweaked to fetch all parameters for stock calculations.