Add sector display option to overview plots
The update adds an option on the overview plot display to show or hide sectors. A new switch input is introduced to allow users to enable or disable the display of sectors in the overview plot. It also includes code that will adjust the plot based on the user's choice to either see all sectors in the chosen scenario or just display a cumulative value.