Add agriculture visualization scripts to dashboard
The code has been updated to include 'agriculture' as a new category in the dashboard. This includes creating callbacks, processing scripts and visualization scripts, mainly for handling data related to the 'agriculture' sector. Some of the existing files are moved to new directories for better code organization and readability. Moreover, the 'ghg.py' has been renamed to 'agriculture.py' to better highlight its function and relevance.